gnl: Properly set operations basetime
For operations (especially mixers) to be able to do proper
synchronization they need the segment from different sources to be in
the same running time context.

This commit makes sure that the basetime of the incoming segments of
operations will be the same, and equal to the duration of the
already "played" time of the operation at the position at which
the new portion of the composition starts.