Properly set operations basetime
For operations (especially mixers) to be able to do proper
synchronization they need the segment from different sources to be in
the same running time context.

This commit make sure that the basetime of the incoming segments of
operations will be the sames, and equal to the duration of the
already "played" time of the operation at the position at which
the the new porsion of the composition starts.